Update on the project
Ethan and I implemented a line of code that prompts the user for their name and stores this data in the "thank you for your service ___!" string displayed on the screen. We also included a list of military branches, but have not yet given users the ability to click on it.